+++ packages/miniprogram-render/src/node/style.js
@@ -59,6 +59,7 @@ class Style {
     $$init(onUpdate) {
         this.$_doUpdate = onUpdate || (() => {})
         this.$_disableCheckUpdate = false // 是否禁止检查更新
+        this.$__vars = null
     }
 
     /**
@@ -67,6 +68,7 @@ class Style {
     $$destroy() {
         this.$_doUpdate = null
         this.$_disableCheckUpdate = false
+        this.$__vars = null
 
         styleList.forEach(name => {
             this[`$_${name}`] = undefined
@@ -88,6 +90,14 @@ class Style {
     }
 
     /**
+     * css 变量存储
+     */
+    get $_vars() {
+        if (!this.$__vars) this.$__vars = {}
+        return this.$__vars
+    }
+
+    /**
      * 检查更新
      */
     $_checkUpdate() {
@@ -100,8 +110,25 @@ class Style {
      * 对外属性和方法
      */
     get cssText() {
-        const joinText = styleList.filter(name => this[`$_${name}`]).map(name => `${tool.toDash(name)}:${this['$_' + name]}`).join(';').trim()
-        return joinText ? `${joinText};` : ''
+        let joinText = styleList
+            .filter(name => this[`$_${name}`])
+            .map(name => `${tool.toDash(name)}:${this['$_' + name]}`)
+            .join(';')
+            .trim()
+        joinText = joinText ? `${joinText};` : ''
+
+        if (this.$__vars) {
+            const vars = this.$_vars
+            const varsText = Object
+                .keys(vars)
+                .filter(name => vars[name])
+                .map(name => `${name}:${vars[name]}`)
+                .join(';')
+                .trim()
+            joinText = varsText ? `${joinText}${varsText};` : joinText
+        }
+
+        return joinText
     }
 
     set cssText(styleText) {
@@ -114,17 +141,26 @@ class Style {
 
         this.$_disableCheckUpdate = true // 将每条规则的设置合并为一次更新
         for (const name of styleList) {
-            this[name] = rules[name]
+            this.setProperty(name, rules[name])
         }
         this.$_disableCheckUpdate = false
         this.$_checkUpdate()
     }
 
+    setProperty(name, value) {
+        if (typeof name !== 'string') return ''
+
+        if (name.indexOf('--') === 0) {
+            this.$_vars[name] = value // css 变量
+            this.$_checkUpdate()
+        } else this[name] = value
+    }
+
     getPropertyValue(name) {
         if (typeof name !== 'string') return ''
 
-        name = tool.toCamel(name)
-        return this[name] || ''
+        if (name.indexOf('--') === 0) return this.$_vars[name] || ''
+        else return this[tool.toCamel(name)] || ''
     }
 }